+++ src/fi/evident/dalesbred/plugin/idea/inspections/DalesbredIncorrectParameterCountInspection.java
@@ -43,6 +43,7 @@ public class DalesbredIncorrectParameterCountInspection extends BaseJavaLocalIns
     private static final PsiMethodCallPattern FIND_METHOD_CALL = psiExpression().methodCall(or(findMethod(), executeQueryMethod()));
     private static final PsiMethodCallPattern SQL_QUERY_METHOD_CALL = psiExpression().methodCall(sqlQueryMethod());
     private static final PsiMethodCallPattern UPDATE_METHOD_CALL = psiExpression().methodCall(updateMethod());
+    private static final PsiMethodCallPattern UPDATE_AND_PROCESS_GENERATED_KEYS_METHOD_CALL = psiExpression().methodCall(updateAndProcessGeneratedKeysMethod());
 
     @NotNull
     @Override
@@ -72,6 +73,9 @@ public class DalesbredIncorrectParameterCountInspection extends BaseJavaLocalIns
                 } else if (UPDATE_METHOD_CALL.accepts(expression)) {
                     PsiExpression[] parameters = expression.getArgumentList().getExpressions();
                     verifyQueryParameterCount(parameters, 0, holder);
+                } else if (UPDATE_AND_PROCESS_GENERATED_KEYS_METHOD_CALL.accepts(expression)) {
+                    PsiExpression[] parameters = expression.getArgumentList().getExpressions();
+                    verifyQueryParameterCount(parameters, 1, holder);
                 }
             }
         };